+++ tests/Unit/LoopbackAdapter.unit.Tests.ps1
@@ -322,6 +322,7 @@ InModuleScope $ProjectName {
             Mock -CommandName Set-NetIPInterface
             Mock -CommandName Get-CimInstance -MockWith { '192.168.0.1' }
             Mock -CommandName Start-Sleep
+            Mock -CommandName Wait-ForDevconUpdate
 
             It 'Should not throw exception' {
                 {
@@ -363,6 +364,7 @@ InModuleScope $ProjectName {
             }
             Mock -CommandName devcon
             Mock -CommandName Get-LoopbackAdapter -MockWith $script:adaptersWithOneLoopbackAdapter_mock
+            Mock -CommandName Wait-ForDevconUpdate
 
             It 'Should throw expected exception' {
                 {
@@ -393,17 +395,13 @@ InModuleScope $ProjectName {
                 -ParameterFilter {
                     $Name -eq 'LoopbackAdapter'
                 }
-            Mock -CommandName Get-NetAdapter `
-                -MockWith @script:adaptersWithOneLoopbackAdapter_mock `
-                -ParameterFilter {
-                    $Name -eq 'LoopbackAdapter'
-                }
             Mock -CommandName Install-Devcon -MockWith {
                 @{
                     FullName = 'devcon'
                 }
             }
             Mock -CommandName devcon
+            Mock -CommandName Wait-ForDevconUpdate
 
             It 'Should not throw exception' {
                 {
@@ -458,6 +456,63 @@ InModuleScope $ProjectName {
 
             It 'Should call choco' {
                 Assert-MockCalled -CommandName Choco
+            }
+        }
+    }
+
+    Describe 'Wait-ForDevconUpdate' -Tag 'Unit' {
+        BeforeAll {
+            Mock -CommandName Get-Process
+            Mock -CommandName Wait-Process
+            Mock -CommandName Start-Sleep
+
+            $script:registryError = (
+                'Illegal operation attempted on a registry key that has been marked for deletion.'
+            )
+        }
+
+        Context 'When the registry is updated right away' {
+            BeforeAll {
+                Mock -CommandName Get-NetAdapter -MockWith {
+                    $script:adaptersWithNoLoopbackAdapter_mock
+                }
+            }
+
+            It 'Should not throw an exception' {
+                { Wait-ForDevconUpdate } | Should -Not -Throw $script:registryError
+            }
+        }
+
+        Context 'When the registry is not updated right away but then updates' {
+            BeforeAll {
+                $script:getNetAdapterCounter = 0
+
+                Mock -CommandName Get-NetAdapter -MockWith {
+                    if ($script:getNetAdapterCounter -eq 0) {
+                        $script:getNetAdapterCounter++
+                        throw [Microsoft.Management.Infrastructure.CimException]::new($script:registryError)
+                    }
+                    else {
+                        $script:adaptersWithNoLoopbackAdapter_mock
+                    }
+                }
+            }
+
+            It 'Should catch the expected error and not throw an exception' {
+                { Wait-ForDevconUpdate } | Should -Not -Throw $script:registryError
+                $script:getNetAdapterCounter | Should -BeGreaterThan 0
+            }
+        }
+
+        Context 'When the registry is not updated before the timeout' {
+            BeforeAll {
+                Mock -CommandName Get-NetAdapter -MockWith {
+                    throw [Microsoft.Management.Infrastructure.CimException]::new($script:registryError)
+                }
+            }
+
+            It 'Should throw an exception' {
+                { Wait-ForDevconUpdate -RegistryUpdateTimeout 1 } | Should -Throw $script:registryError
             }
         }
     }